realtek: add demo support for the Zyxel XGS1210-12
The Zyxel XGS1210-12 is a 8x1GBit, 2x2.5GBit and 2x10GBit SFP+
switch. The device has the following hardware:
    - RTL9302B SoC
    - Macronix MX25L12833F (16MB flash)
    - Nanja NT5CC64M16GP-1 (128MB DDR3 SDRAM)
    - RTL8231 GPIO extender to control the port LEDs
    - RTL8218D 8x Gigabit PHY
    - 2 RTL8226 2.5 Gigabit PHYs
The 2 SFP+ uplink ports are not supported by the realtek switch driver, yet.
The bi-color and tri-color (for the 2.5 Gig ports) LEDs are driven directly by
the SoC. A reset button is present on the front but not supported.
Behind the air-vents on the right a 3.3V uart header is externally accessible.
Pins are from top to bottom Vcc(3.3V), TX, RX and GND. Serial connection is via
115200 baud.

To install, boot into u-boot, halt via pressing <ESC>, load initramfs via tftp
and boot via bootm.
Installation to flash was not tested.

Signed-off-by: Birger Koblitz <git@birger-koblitz.de>